+++ src/backend/InvenTree/InvenTree/api_version.py
@@ -1,14 +1,37 @@
 """InvenTree API version information."""
 
 # InvenTree API version
-INVENTREE_API_VERSION = 475
+INVENTREE_API_VERSION = 482
 """Increment this API version number whenever there is a significant change to the API that any clients need to know about."""
 
 INVENTREE_API_TEXT = """
 
-v475 -> 2026-03-15 : https://github.com/inventree/InvenTree/pull/11540
+v482 -> 2026-03-15 : https://github.com/inventree/InvenTree/pull/11540
     - Add id to the ordering fields of the Parts model
-    
+
+v481 -> 2026-04-28 : https://github.com/inventree/InvenTree/pull/11825
+    - Adds new "bom" ruleset and associated permissions for BOM management, separate from the "part" ruleset which remains focused on part management
+
+v480 -> 2026-04-27 : https://github.com/inventree/InvenTree/pull/11816
+    - The "issued_by" field on the Build API endpoint is now read-only, and is automatically set to the current user when a build is created
+
+v479 -> 2026-04-11 : https://github.com/inventree/InvenTree/pull/11723
+    - POST /api/notifications/readall/ now requires a POST action
+    - POST /api/admin/email/test/ - now returns a 200 on. a successful test
+    - GET /api/notifications/ - now uses user-centric permissions, not a general read
+
+v478 -> 2026-04-11 : https://github.com/inventree/InvenTree/pull/11073
+    - Add OptionalField class for cleaner handling of optional fields in serializers
+
+v477 -> 2026-04-11 : https://github.com/inventree/InvenTree/pull/11617
+    - Non-functional refactor, adaptations of descriptions
+
+v476 -> 2026-04-09 : https://github.com/inventree/InvenTree/pull/11705
+    - Adds sorting / filtering / searching functionality to the SelectionListEntry API endpoint
+
+v475 -> 2026-04-09 : https://github.com/inventree/InvenTree/pull/11702
+    - Adds "updated" and "updated_by" fields to the LabelTemplate and ReportTemplate API endpoints
+
 v474 -> 2026-04-08 : https://github.com/inventree/InvenTree/pull/11693
     - Adds DataImportMixin to the ManufacturerPartList API endpoint