+++ src/backend/InvenTree/part/models.py
@@ -2058,7 +2058,12 @@ class Part(
 
         Note: Does *NOT* delete inherited BOM items!
         """
+        import part.tasks as part_tasks
+
         self.bom_items.all().delete()
+
+        # Offload task to re-validate the BOM for this assembly
+        InvenTree.tasks.offload_task(part_tasks.check_bom_valid, self.pk, group='part')
 
     def getRequiredParts(self, recursive=False, parts=None):
         """Return a list of parts required to make this part (i.e. BOM items).
@@ -2473,6 +2478,7 @@ class Part(
             category__in=categories
         ).order_by('-category__level')
 
+        template_ids = set()
         parameters = []
         content_type = ContentType.objects.get_for_model(Part)
 
@@ -2483,6 +2489,12 @@ class Part(
             ).exists():
                 continue
 
+            # Ensure we do not create duplicate parameters if multiple categories have the same template
+            if category_template.template.pk in template_ids:
+                continue
+
+            template_ids.add(category_template.template.pk)
+
             parameters.append(
                 Parameter(
                     template=category_template.template,
@@ -2492,8 +2504,7 @@ class Part(
                 )
             )
 
-        if len(parameters) > 0:
-            Parameter.objects.bulk_create(parameters)
+        Parameter.objects.bulk_create(parameters)
 
     def getTestTemplates(
         self, required=None, include_parent: bool = True, enabled=None
@@ -2625,9 +2636,7 @@ class Part(
         parts = []
 
         # Child parts
-        children = self.get_descendants(include_self=False)
-
-        for child in children:
+        for child in self.get_descendants(include_self=False):
             parts.append(child)
 
         # Immediate parent, and siblings
@@ -3923,8 +3932,11 @@ class BomItem(InvenTree.models.MetadataMixin, InvenTree.models.InvenTreeModel):
             assemblies = set()
 
             if db_instance:
-                # Find all assemblies which use this BomItem *after* we save
+                # Find all assemblies which use this BomItem *before* we save
                 assemblies.update(db_instance.get_assemblies())
+
+            # Update the set of assemblies to include those which use this BomItem *after* we save
+            assemblies.update(self.get_assemblies())
 
             for assembly in assemblies:
                 # Offload task to update the checksum for this assembly
@@ -4079,7 +4091,9 @@ class BomItem(InvenTree.models.MetadataMixin, InvenTree.models.InvenTreeModel):
         These fields are used to calculate the checksum hash of this BOM item.
         """
         return [
+            'part',
             'part_id',
+            'sub_part',
             'sub_part_id',
             'quantity',
             'setup_quantity',