+++ src/frontend/src/forms/CommonForms.tsx
@@ -1,8 +1,8 @@
 import { IconUsers } from '@tabler/icons-react';
-import { useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 
 import { ApiEndpoints } from '@lib/enums/ApiEndpoints';
-import type { ModelType } from '@lib/enums/ModelType';
+import { ModelType } from '@lib/enums/ModelType';
 import { apiUrl } from '@lib/functions/Api';
 import type { ApiFormFieldSet } from '@lib/types/Forms';
 import { t } from '@lingui/core/macro';
@@ -12,6 +12,7 @@ import type {
 } from '../components/render/StatusRenderer';
 import { useApi } from '../contexts/ApiContext';
 import { useGlobalStatusState } from '../states/GlobalStatusState';
+import { useUserState } from '../states/UserState';
 
 export function projectCodeFields(): ApiFormFieldSet {
   return {
@@ -97,6 +98,25 @@ export function extraLineItemFields(): ApiFormFieldSet {
   };
 }
 
+export function useParameterTemplateFields(): ApiFormFieldSet {
+  return useMemo(() => {
+    return {
+      name: {},
+      description: {},
+      units: {},
+      model_type: {},
+      choices: {},
+      checkbox: {},
+      selectionlist: {
+        filters: {
+          active: true
+        }
+      },
+      enabled: {}
+    };
+  }, []);
+}
+
 export function useParameterFields({
   modelType,
   modelId
@@ -106,18 +126,49 @@ export function useParameterFields({
 }): ApiFormFieldSet {
   const api = useApi();
 
+  const user = useUserState.getState();
+
+  const templateCreateFields = useParameterTemplateFields();
+
+  const [selectionListId, setSelectionListId] = useState<number | null>(null);
+
   // Valid field choices
   const [choices, setChoices] = useState<any[]>([]);
 
   // Field type for "data" input
-  const [fieldType, setFieldType] = useState<'string' | 'boolean' | 'choice'>(
-    'string'
-  );
+  const [fieldType, setFieldType] = useState<
+    'string' | 'boolean' | 'choice' | 'related field'
+  >('string');
 
+  // Memoized value for the "data" field
   const [data, setData] = useState<string>('');
+
+  const fetchSelectionEntry = useCallback(
+    (value: any) => {
+      if (!value || !selectionListId) {
+        return null;
+      }
+
+      return api
+        .get(apiUrl(ApiEndpoints.selectionentry_list, selectionListId), {
+          params: {
+            value: value
+          }
+        })
+        .then((response) => {
+          if (response.data && response.data.length == 1) {
+            return response.data[0];
+          } else {
+            return null;
+          }
+        });
+    },
+    [selectionListId]
+  );
 
   // Reset the field type and choices when the model changes
   useEffect(() => {
+    setSelectionListId(null);
     setFieldType('string');
     setChoices([]);
     setData('');
@@ -139,6 +190,8 @@ export function useParameterFields({
           enabled: true
         },
         onValueChange: (value: any, record: any) => {
+          setSelectionListId(record?.selectionlist || null);
+
           // Adjust the type of the "data" field based on the selected template
           if (record?.checkbox) {
             // This is a "checkbox" field
@@ -162,36 +215,43 @@ export function useParameterFields({
               setFieldType('string');
             }
           } else if (record?.selectionlist) {
-            api
-              .get(
-                apiUrl(ApiEndpoints.selectionlist_detail, record.selectionlist)
-              )
-              .then((res) => {
-                setChoices(
-                  res.data.choices.map((item: any) => {
-                    return {
-                      value: item.value,
-                      display_name: item.label
-                    };
-                  })
-                );
-                setFieldType('choice');
-              });
-          } else {
-            setChoices([]);
-            setFieldType('string');
+            setFieldType('related field');
           }
-        }
+        },
+        addCreateFields: user.isStaff() ? templateCreateFields : undefined
       },
       data: {
         value: data,
-        onValueChange: (value: any) => {
-          setData(value);
+        onValueChange: (value: any, record: any) => {
+          if (fieldType === 'related field' && selectionListId) {
+            // For related fields, we need to store the selected primary key value (not the string representation)
+            setData(record?.value ?? value);
+          } else {
+            setData(value);
+          }
         },
         type: fieldType,
         field_type: fieldType,
         choices: fieldType === 'choice' ? choices : undefined,
         default: fieldType === 'boolean' ? false : undefined,
+        pk_field:
+          fieldType === 'related field' && selectionListId
+            ? 'value'
+            : undefined,
+        model:
+          fieldType === 'related field' && selectionListId
+            ? ModelType.selectionentry
+            : undefined,
+        api_url:
+          fieldType === 'related field' && selectionListId
+            ? apiUrl(ApiEndpoints.selectionentry_list, selectionListId)
+            : undefined,
+        filters:
+          fieldType === 'related field'
+            ? {
+                active: true
+              }
+            : undefined,
         adjustValue: (value: any) => {
           // Coerce boolean value into a string (required by backend)
 
@@ -204,9 +264,19 @@ export function useParameterFields({
           }
 
           return v;
-        }
+        },
+        singleFetchFunction: fetchSelectionEntry
       },
       note: {}
     };
-  }, [data, modelType, fieldType, choices, modelId]);
+  }, [
+    data,
+    modelType,
+    fieldType,
+    choices,
+    modelId,
+    selectionListId,
+    templateCreateFields,
+    user
+  ]);
 }