+++ src/frontend/src/tables/general/ParameterTemplateTable.tsx
@@ -1,3 +1,4 @@
+import useTable from '@lib/hooks/UseTable';
 import {
   AddItemButton,
   ApiEndpoints,
@@ -12,13 +13,13 @@ import type { TableFilter } from '@lib/types/Filters';
 import type { RowAction, TableColumn } from '@lib/types/Tables';
 import { t } from '@lingui/core/macro';
 import { useCallback, useMemo, useState } from 'react';
+import { useParameterTemplateFields } from '../../forms/CommonForms';
 import { useFilters } from '../../hooks/UseFilter';
 import {
   useCreateApiFormModal,
   useDeleteApiFormModal,
   useEditApiFormModal
 } from '../../hooks/UseForm';
-import { useTable } from '../../hooks/UseTable';
 import { useUserState } from '../../states/UserState';
 import { BooleanColumn, DescriptionColumn } from '../ColumnRenderers';
 import { InvenTreeTable } from '../InvenTreeTable';
@@ -30,18 +31,7 @@ export default function ParameterTemplateTable() {
   const table = useTable('parameter-templates');
   const user = useUserState();
 
-  const parameterTemplateFields: ApiFormFieldSet = useMemo(() => {
-    return {
-      name: {},
-      description: {},
-      units: {},
-      model_type: {},
-      choices: {},
-      checkbox: {},
-      selectionlist: {},
-      enabled: {}
-    };
-  }, []);
+  const parameterTemplateFields: ApiFormFieldSet = useParameterTemplateFields();
 
   const newTemplate = useCreateApiFormModal({
     url: ApiEndpoints.parameter_template_list,